Added a preprocessor hook for the SDMX transform
When the core data model has changed, we need a way to transform the SDMX document from the previous model to the new model for the backward compatibility. This commit adds `to_sdmx` method to `DataModelSerializers` class so that a preprocessor for the SDMX transform can be added per model basis.